Split out util getsizeof helper function tests
